WIP : Entities improvements
