+++ src/lib/kotlin/slatekit-entities/src/main/kotlin/slatekit/entities/core/ServiceSupport.kt
@@ -2,6 +2,7 @@ package slatekit.entities.core
 
 import slatekit.entities.Entity
 import slatekit.entities.EntityRepo
+import java.util.*
 
 interface ServiceSupport<TId, T> where TId : Comparable<TId>, T : Entity<TId> {
 
@@ -16,5 +17,16 @@ interface ServiceSupport<TId, T> where TId : Comparable<TId>, T : Entity<TId> {
      */
     fun applyFieldData(mode: EntityAction, entity: T): T {
         return entity
+    }
+
+
+    fun isCreated(id: TId): Boolean {
+        return when (id) {
+            is Int -> id > 0
+            is Long -> id > 0L
+            is String -> !id.isEmpty()
+            is UUID -> !id.toString().trim().isEmpty()
+            else -> false
+        }
     }
 }
