+++ src/lib/kotlin/slatekit-entities/src/main/kotlin/slatekit/entities/features/EntitySaves.kt
@@ -5,6 +5,8 @@ import slatekit.entities.Entity
 import slatekit.entities.core.EntityAction
 import slatekit.entities.core.EntityEvent
 import slatekit.entities.core.ServiceSupport
+import slatekit.results.Try
+import slatekit.results.builders.Tries
 
 interface EntitySaves<TId, T> : ServiceSupport<TId, T> where TId: kotlin.Comparable<TId>, T: Entity<TId> {
 
@@ -13,17 +15,35 @@ interface EntitySaves<TId, T> : ServiceSupport<TId, T> where TId: kotlin.Compara
      * checking its persisted flag.
      * @param entity
      */
-    fun save(entity: T?) {
-        entity?.let { item ->
-            val finalEntity = applyFieldData(EntityAction.EntitySave, item)
-            repoT().save(finalEntity)
-
-            // Event out
-            if ( this is EntityHooks) {
-                this.onEntityEvent(EntityEvent.EntitySaved(item.identity(), item, DateTime.now()))
-            }
+    fun save(entity: T?): Try<TId> {
+        val result = try {
+            entity?.let { item ->
+                val saveResult = when (item.isPersisted()) {
+                    false -> {
+                        val finalEntity = applyFieldData(EntityAction.EntityCreate, item)
+                        val id = repoT().create(finalEntity)
+                        if (isCreated(id)) Tries.success(id) else Tries.errored("Error creating item")
+                    }
+                    true -> {
+                        val finalEntity = applyFieldData(EntityAction.EntityUpdate, item)
+                        val updated = repoT().update(finalEntity)
+                        if (updated) Tries.success(finalEntity.identity()) else Tries.errored("Error updating item")
+                    }
+                }
+                // Event out
+                saveResult.onSuccess {
+                    if (this is EntityHooks) {
+                        this.onEntityEvent(EntityEvent.EntitySaved(item.identity(), item, DateTime.now()))
+                    }
+                }
+                saveResult
+            } ?: Tries.errored("Entity not provided")
+        } catch(ex:Exception) {
+            Tries.errored<TId>(ex)
         }
+        return result
     }
+
 
     /**
      * saves all the entities
